set neo4j tests in one class
